+++ doc/src/release_notes.rst
@@ -51,6 +51,10 @@ To see the full list of changes check the list of `Git commits <https://github.c
 pgRouting 3.0.0 Release Notes
 -------------------------------------------------------------------------------
 
+.. rubric:: Fixed Issues
+
+* `#1153 <https://github.com/pgRouting/pgrouting/issues/1153>`__: Renamed pgr_eucledianTSP to pgr_TSPeuclidean
+
 .. rubric:: New Experimental functions
 
 * Prim family
@@ -145,6 +149,8 @@ pgRouting 3.0.0 Release Notes
     * pgr_maxFlowBoykovKolmogorov
     * pgr_maxFlowEdmondsKarp
     * pgr_maximumcardinalitymatching
+
+* TSP old signatures